Backend/requirements.txt
# FastAPI e dependências web
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Banco de dados
sqlalchemy==1.4.54
pyodbc==5.2.0  # Para conexão com SQL Server
ollama>=0.3.0,<0.4  # Cliente Python para Ollama
python-docx==1.1.2  # Processamento de arquivos DOCX

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# IA e Machine Learning
langchain==0.3.9
langchain-ollama==0.2.1
langchain-chroma==0.1.4
langchain-community==0.3.8

# Web scraping e parsing (NOVAS DEPENDÊNCIAS)
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# Utilidades
pydantic==2.10.3
